gui/requirements.txt
django
python-decouple
pygame
tensorflow
numpy
neat-python
whitenoise
gunicorn